+++ include/votca/xtp/calculatorfactory.h
@@ -41,28 +41,12 @@ class Calculatorfactory
  public:
   static void RegisterAll(void);
 
-  /**
-     Create an instance of the object identified by key.
-  *  Overwritten to load calculator defaults
-  */
-  QMCalculator *Create(const std::string &key);
-
   friend Calculatorfactory &Calculators();
 };
 
 inline Calculatorfactory &Calculators() {
   static Calculatorfactory instance;
   return instance;
-}
-
-inline QMCalculator *Calculatorfactory::Create(const std::string &key) {
-  assoc_map::const_iterator it(getObjects().find(key));
-  if (it != getObjects().end()) {
-    QMCalculator *calc = (it->second)();
-    return calc;
-  } else {
-    throw std::runtime_error("factory key " + key + " not found.");
-  }
 }
 
 }  // namespace xtp