+++ src/libxtp/tools/molpol.cc
@@ -74,8 +74,8 @@ void MolPol::ParseOptions(const tools::Property& options) {
     XTP_LOG(Log::error, log)
         << "Using package <" << qm_package << ">" << std::flush;
     QMPackageFactory::RegisterAll();
-    std::unique_ptr<QMPackage> qmpack =
-        std::unique_ptr<QMPackage>(QMPackages().Create(qm_package));
+    std::unique_ptr<QMPackage> qmpack = std::unique_ptr<QMPackage>(
+        QMPackageFactory::QMPackages().Create(qm_package));
     qmpack->setLog(&log);
     qmpack->setRunDir(".");
     qmpack->setLogFileName(log_file);