+++ src/tests/test_orca.cc
@@ -40,7 +40,11 @@ BOOST_AUTO_TEST_CASE(polar_test) {
 
   QMPackageFactory::RegisterAll();
   std::unique_ptr<QMPackage> orca =
-      std::unique_ptr<QMPackage>(QMPackages().Create("orca"));
+      QMPackageFactory::QMPackages().Create("orca");
+  auto keys = QMPackageFactory::QMPackages().getKeys();
+  for (auto key : keys) {
+    std::cout << key << std::endl;
+  }
   Logger log;
   orca->setLog(&log);
   orca->setRunDir(std::string(XTP_TEST_DATA_FOLDER) + "/orca");
@@ -63,7 +67,7 @@ BOOST_AUTO_TEST_CASE(ext_charges_test) {
   libint2::initialize();
   QMPackageFactory::RegisterAll();
   std::unique_ptr<QMPackage> orca =
-      std::unique_ptr<QMPackage>(QMPackages().Create("orca"));
+      QMPackageFactory::QMPackages().Create("orca");
   Logger log;
   orca->setLog(&log);
   orca->setRunDir(std::string(XTP_TEST_DATA_FOLDER) + "/orca");
@@ -151,7 +155,7 @@ BOOST_AUTO_TEST_CASE(charges_test) {
   libint2::initialize();
   QMPackageFactory::RegisterAll();
   std::unique_ptr<QMPackage> orca =
-      std::unique_ptr<QMPackage>(QMPackages().Create("orca"));
+      QMPackageFactory::QMPackages().Create("orca");
   Logger log;
   orca->setLog(&log);
   orca->setRunDir(std::string(XTP_TEST_DATA_FOLDER) + "/orca");
@@ -195,7 +199,7 @@ BOOST_AUTO_TEST_CASE(opt_test) {
 
   QMPackageFactory::RegisterAll();
   std::unique_ptr<QMPackage> orca =
-      std::unique_ptr<QMPackage>(QMPackages().Create("orca"));
+      QMPackageFactory::QMPackages().Create("orca");
   Logger log;
   orca->setLog(&log);
   orca->setRunDir(std::string(XTP_TEST_DATA_FOLDER) + "/orca");
@@ -260,7 +264,7 @@ BOOST_AUTO_TEST_CASE(input_generation_version_4_0_1) {
 
   QMPackageFactory::RegisterAll();
   std::unique_ptr<QMPackage> orca =
-      std::unique_ptr<QMPackage>(QMPackages().Create("orca"));
+      QMPackageFactory::QMPackages().Create("orca");
   Logger log;
   orca->setLog(&log);
   orca->setRunDir(".");